Refactor join flow
